feat(floor): keep a minimum amount of space between vertical delimitations
- Keep a mimimum amount of space between vertical delimitations because most generations just looked like shit.
- Add an option to change that space value.

Before, only a few generations looked good. Most generations had a few vertical delimitations that were so close, it didn't look good.
Now, you can play around with that space value too match what you prefer.
The higher the value, the more it looks bad but the more difference there are.
The lower the value, the more it looks good but the more they look all the same.
4 to 8 are pretty good values.